TFP-6470 Jackson3 som default, bruk i rest/graphql
