+++ felles/mapper/src/main/java/no/nav/vedtak/mapper/json/DefaultJsonMapper.java
@@ -1,7 +1,6 @@
 package no.nav.vedtak.mapper.json;
 
 import java.io.File;
-import java.io.IOException;
 import java.io.InputStream;
 import java.util.List;
 import java.util.Map;
@@ -9,17 +8,15 @@ import java.util.TimeZone;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.PropertyAccessor;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.MapperFeature;
-import com.fasterxml.jackson.databind.SerializationFeature;
-import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
 import no.nav.vedtak.exception.TekniskException;
+import tools.jackson.core.JacksonException;
+import tools.jackson.core.type.TypeReference;
+import tools.jackson.databind.DeserializationFeature;
+import tools.jackson.databind.JsonNode;
+import tools.jackson.databind.MapperFeature;
+import tools.jackson.databind.cfg.EnumFeature;
+import tools.jackson.databind.json.JsonMapper;
 
 /**
  * Se også jackson-core, pakke json, JsonReadFeature og JsonWriteFeature for mer Java/Json-nær konfig/features
@@ -30,7 +27,7 @@ public class DefaultJsonMapper {
 
     }
 
-    private static final JsonMapper MAPPER = createJsonMapper();
+    private static final JsonMapper MAPPER = createJsonMapperBuilder().build();
 
     // Foretrekker denne - men bruk heller metoder nedenfor direkte enn å assigne til lokale variable
     public static JsonMapper getJsonMapper() {
@@ -38,34 +35,32 @@ public class DefaultJsonMapper {
     }
 
     // Bruk denne for kun for ContextResolver (JacksonJsonConfig-klasser) som skal legge til (de)serializers eller registrere subtypes
-    public static JsonMapper getCopyFromDefaultJsonMapper() {
-        return MAPPER.copy();
+    public static JsonMapper.Builder getCopyFromDefaultJsonMapper() {
+        return createJsonMapperBuilder();
     }
 
-    private static JsonMapper createJsonMapper() {
+    private static JsonMapper.Builder createJsonMapperBuilder() {
         return JsonMapper.builder()
-            .addModule(new Jdk8Module())
-            .addModule(new JavaTimeModule())
             .defaultTimeZone(TimeZone.getTimeZone("Europe/Oslo"))
-            .disable(SerializationFeature.FAIL_ON_EMPTY_BEANS)
-            .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-            .disable(SerializationFeature.WRITE_DURATIONS_AS_TIMESTAMPS)
-            .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
-            .enable(DeserializationFeature.READ_UNKNOWN_ENUM_VALUES_USING_DEFAULT_VALUE)
+            .disable(DeserializationFeature.FAIL_ON_NULL_FOR_PRIMITIVES) // Var noen tester med null for booleans
+            .enable(EnumFeature.READ_UNKNOWN_ENUM_VALUES_USING_DEFAULT_VALUE)
             .enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_PROPERTIES) // TODO: Trengs denne? Sak har kjørt lenge uten
-            .defaultPropertyInclusion(JsonInclude.Value.construct(JsonInclude.Include.NON_NULL, JsonInclude.Include.NON_NULL))
-            .visibility(PropertyAccessor.GETTER, JsonAutoDetect.Visibility.NONE)
-            .visibility(PropertyAccessor.SETTER, JsonAutoDetect.Visibility.NONE)
-            .visibility(PropertyAccessor.IS_GETTER, JsonAutoDetect.Visibility.NONE)
-            .visibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY)
-            .visibility(PropertyAccessor.CREATOR, JsonAutoDetect.Visibility.ANY)
-            .build();
+            .changeDefaultPropertyInclusion(a -> a
+                .withValueInclusion(JsonInclude.Include.NON_NULL)
+                .withContentInclusion(JsonInclude.Include.NON_NULL))
+            .changeDefaultVisibility(v -> v
+                    .withGetterVisibility(JsonAutoDetect.Visibility.NONE)
+                    .withIsGetterVisibility(JsonAutoDetect.Visibility.NONE)
+                    .withSetterVisibility(JsonAutoDetect.Visibility.NONE)
+                    .withFieldVisibility(JsonAutoDetect.Visibility.ANY)
+                    .withCreatorVisibility(JsonAutoDetect.Visibility.ANY)
+                    .withScalarConstructorVisibility(JsonAutoDetect.Visibility.ANY));
     }
 
     public static <T> T fromJson(String json, TypeReference<T> typeReference) {
         try {
             return MAPPER.readerFor(typeReference).readValue(json);
-        } catch (IOException e) {
+        } catch (JacksonException e) {
             throw deserializationException(e);
         }
     }
@@ -73,7 +68,15 @@ public class DefaultJsonMapper {
     public static <T> T fromJson(String json, Class<T> clazz) {
         try {
             return MAPPER.readerFor(clazz).readValue(json);
-        } catch (IOException e) {
+        } catch (JacksonException e) {
+            throw deserializationException(e);
+        }
+    }
+
+    public static <T> T fromJson(File json, Class<T> clazz) {
+        try {
+            return MAPPER.readerFor(clazz).readValue(json);
+        } catch (JacksonException e) {
             throw deserializationException(e);
         }
     }
@@ -81,15 +84,7 @@ public class DefaultJsonMapper {
     public static <T> T fromJson(InputStream json, Class<T> clazz) {
         try {
             return MAPPER.readerFor(clazz).readValue(json);
-        } catch (IOException e) {
-            throw deserializationException(e);
-        }
-    }
-
-    public static <T> T fromJson(File json, Class<T> clazz) {
-        try {
-            return MAPPER.readerFor(clazz).readValue(json, clazz);
-        } catch (IOException e) {
+        } catch (JacksonException e) {
             throw deserializationException(e);
         }
     }
@@ -97,7 +92,7 @@ public class DefaultJsonMapper {
     public static <T> List<T> listFromJson(String json, Class<T> clazz) {
         try {
             return MAPPER.readerForListOf(clazz).readValue(json);
-        } catch (IOException e) {
+        } catch (JacksonException e) {
             throw deserializationException(e);
         }
     }
@@ -105,7 +100,7 @@ public class DefaultJsonMapper {
     public static <T> Map<String, T> mapFromJson(String json, Class<T> clazz) {
         try {
             return MAPPER.readerForMapOf(clazz).readValue(json);
-        } catch (IOException e) {
+        } catch (JacksonException e) {
             throw deserializationException(e);
         }
     }
@@ -113,7 +108,7 @@ public class DefaultJsonMapper {
     public static JsonNode treeFromJson(String json) {
         try {
             return MAPPER.readTree(json);
-        } catch (IOException e) {
+        } catch (JacksonException e) {
             throw deserializationException(e);
         }
     }
@@ -121,7 +116,7 @@ public class DefaultJsonMapper {
     public static String toJson(Object obj) {
         try {
             return MAPPER.writeValueAsString(obj);
-        } catch (IOException e) {
+        } catch (JacksonException e) {
             throw serializationException(e);
         }
     }
@@ -129,16 +124,16 @@ public class DefaultJsonMapper {
     public static String toPrettyJson(Object obj) {
         try {
             return MAPPER.writerWithDefaultPrettyPrinter().writeValueAsString(obj);
-        } catch (IOException e) {
+        } catch (JacksonException e) {
             throw serializationException(e);
         }
     }
 
-    private static TekniskException deserializationException(IOException e) {
-        return new TekniskException("FP-713328", "Fikk IO exception ved deserialisering av JSON", e);
+    private static TekniskException deserializationException(JacksonException e) {
+        return new TekniskException("FP-713328", "Fikk JacksonException ved deserialisering av JSON", e);
     }
 
-    private static TekniskException serializationException(IOException e) {
+    private static TekniskException serializationException(JacksonException e) {
         return new TekniskException("F-208314", "Kunne ikke serialisere objekt til JSON", e);
     }
 }
