+++ felles/server/src/main/java/no/nav/vedtak/server/rest/jackson/JacksonContextResolver.java
@@ -2,12 +2,12 @@ package no.nav.vedtak.server.rest.jackson;
 
 import jakarta.ws.rs.ext.ContextResolver;
 
-import no.nav.vedtak.mapper.json.DefaultJson3Mapper;
+import no.nav.vedtak.mapper.json.DefaultJsonMapper;
 import tools.jackson.databind.json.JsonMapper;
 
-public class Jackson3ContextResolver implements ContextResolver<JsonMapper> {
+public class JacksonContextResolver implements ContextResolver<JsonMapper> {
 
-    private static final JsonMapper MAPPER = DefaultJson3Mapper.getJsonMapper();
+    private static final JsonMapper MAPPER = DefaultJsonMapper.getJsonMapper();
 
     @Override
     public JsonMapper getContext(Class<?> type) {
