+++ felles/server/src/main/java/no/nav/vedtak/server/rest/jackson/JacksonExceptionMapper.java
@@ -8,7 +8,7 @@ import no.nav.vedtak.log.util.LoggerUtils;
 import no.nav.vedtak.server.rest.RestServerFeilUtils;
 import tools.jackson.core.JacksonException;
 
-public class Jackson3ExceptionMapper implements ExceptionMapper<JacksonException> {
+public class JacksonExceptionMapper implements ExceptionMapper<JacksonException> {
 
     @Override
     public Response toResponse(JacksonException exception) {
