+++ felles/server/src/main/java/no/nav/vedtak/server/rest/jackson/JacksonRestFeature.java
@@ -4,13 +4,13 @@ import jakarta.ws.rs.core.Feature;
 import jakarta.ws.rs.core.FeatureContext;
 
 
-public class Jackson3RestFeature implements Feature {
+public class JacksonRestFeature implements Feature {
 
     @Override
     public boolean configure(final FeatureContext context) {
-        context.register(Jackson3ProviderFeature.class);
-        context.register(Jackson3ContextResolver.class);
-        context.register(Jackson3ExceptionMapper.class);
+        context.register(JacksonProviderFeature.class);
+        context.register(JacksonContextResolver.class);
+        context.register(JacksonExceptionMapper.class);
 
         return true;
     }
