+++ felles/server/src/test/java/no/nav/vedtak/server/rest/jackson/JacksonExceptionMapperTest.java
@@ -6,14 +6,13 @@ import java.time.LocalDateTime;
 
 import org.junit.jupiter.api.Test;
 
-import com.fasterxml.jackson.core.JacksonException;
-
 import no.nav.vedtak.feil.FeilDto;
 import no.nav.vedtak.mapper.json.DefaultJsonMapper;
+import tools.jackson.core.JacksonException;
 import tools.jackson.core.exc.UnexpectedEndOfInputException;
 import tools.jackson.databind.exc.InvalidTypeIdException;
 
-class Jackson3ExceptionMapperTest {
+class JacksonExceptionMapperTest {
 
     record A(LocalDateTime tidspunkt) {}
 
@@ -22,7 +21,7 @@ class Jackson3ExceptionMapperTest {
         try {
             DefaultJsonMapper.getJsonMapper().readerFor(FeilDto.class).readValue("{ \"feilkoder\": } \"123\" }");
         } catch (JacksonException e) {
-            try (var resultat = new Jackson2ExceptionMapper().toResponse(e)) {
+            try (var resultat = new JacksonExceptionMapper().toResponse(e)) {
                 var dto = (FeilDto) resultat.getEntity();
                 assertThat(dto.feilmelding()).contains("FP-252294 JSON-feil: Unexpected character ");
             }
@@ -34,7 +33,7 @@ class Jackson3ExceptionMapperTest {
         try {
             DefaultJsonMapper.getJsonMapper().readerFor(A.class).readValue("{ \"tidspunkt\": 123 }");
         } catch (JacksonException e) {
-            try (var resultat = new Jackson2ExceptionMapper().toResponse(e)) {
+            try (var resultat = new JacksonExceptionMapper().toResponse(e)) {
                 var dto = (FeilDto) resultat.getEntity();
                 assertThat(dto.feilmelding()).contains("FP-252294 JSON-feil: raw timestamp");
             }
@@ -43,7 +42,7 @@ class Jackson3ExceptionMapperTest {
 
     @Test
     void skal_mappe_InvalidTypeIdException() {
-        try (var resultat = new Jackson3ExceptionMapper().toResponse(new InvalidTypeIdException(null, "Ukjent type-kode", null, "23525"))) {
+        try (var resultat = new JacksonExceptionMapper().toResponse(new InvalidTypeIdException(null, "Ukjent type-kode", null, "23525"))) {
             var dto = (FeilDto) resultat.getEntity();
             assertThat(dto.feilmelding()).contains("FP-252294 JSON-feil");
         }
@@ -52,7 +51,7 @@ class Jackson3ExceptionMapperTest {
     @Test
     void skal_parse_JsonEOFException(){
         var feilTekst = "Ukjent EOF";
-        try (var resultat = new Jackson3ExceptionMapper().toResponse(new UnexpectedEndOfInputException(null, null, feilTekst))) {
+        try (var resultat = new JacksonExceptionMapper().toResponse(new UnexpectedEndOfInputException(null, null, feilTekst))) {
             var dto = (FeilDto) resultat.getEntity();
 
             assertThat(dto.feilmelding()).contains("FP-252294 JSON-feil: " + feilTekst);
