+++ integrasjon/person-klient/src/main/resources/pdl/pdl.graphqls
@@ -20,7 +20,7 @@ type Query {
     hentGeografiskTilknytning(ident: ID!): GeografiskTilknytning
     hentGeografiskTilknytningBolk(identer: [ID!]!): [hentGeografiskTilknytningBolkResult!]!
     sokPerson(criteria:[Criterion], paging:Paging): PersonSearchResult
-    sokAdresse(criteria:[Criterion], paging:Paging): AdresseSearchResult
+    sokAdresse(criteria:[Criterion], paging:Paging, aggregations:[Aggregation]): AdresseSearchResult
     forslagAdresse(parameters:CompletionParameters): AdresseCompletionResult
     hentAdresse(matrikkelId: ID!): KartverketAdresse
 }
@@ -138,7 +138,6 @@ type Person {
     doedfoedtBarn: [DoedfoedtBarn!]!
     doedsfall: [Doedsfall!]!
     falskIdentitet: FalskIdentitet
-    foedsel: [Foedsel!]! @deprecated(reason: "Erstattet av foedselsdato & foedested")
     foedselsdato: [Foedselsdato!]!
     foedested: [Foedested!]!
     navspersonidentifikator(historikk: Boolean = false): [Navspersonidentifikator!]!
@@ -373,16 +372,6 @@ enum AdressebeskyttelseGradering {
     UGRADERT
 }
 
-type Foedsel {
-    foedselsaar: Int
-    foedselsdato: Date
-    foedeland: String
-    foedested: String
-    foedekommune: String
-    folkeregistermetadata: Folkeregistermetadata
-    metadata: Metadata!
-}
-
 type Foedested {
     foedeland: String
     foedested: String
@@ -542,6 +531,8 @@ type Folkeregisteridentifikator {
 }
 type Navspersonidentifikator {
     identifikasjonsnummer: String!
+    gyldigFraOgMed: Date #TODO Sett som required etter oppdatering i databasen
+    gyldigTilOgMed: Date
     metadata: Metadata!
 }
 
@@ -732,6 +723,10 @@ input Criterion {
     searchHistorical:Boolean
 }
 
+input Aggregation {
+    fieldName:String
+    maxResults:Int = 10
+}
 
 input SearchRule {
     # Sjekker om feltet finnes / at det ikke har en null verdi.
@@ -843,6 +838,7 @@ type SearchMatch {
 }
 type AdresseSearchResult {
     hits : [AdresseSearchHit!]!
+    aggregations: [AggregationResult!]
     pageNumber:Int,
     totalPages:Int
     totalHits:Int
@@ -853,6 +849,16 @@ type AdresseSearchHit {
     matrikkeladresse : MatrikkeladresseResult
     score :Float
 }
+
+type AggregationResult {
+    fieldName:String
+    values: [AggregationValue!]
+}
+type AggregationValue {
+    value:String
+    count:Long
+}
+
 type VegadresseResult {
     matrikkelId:String
     husnummer:Int
