+++ integrasjon/person-klient/src/test/java/no/nav/vedtak/felles/integrasjon/person/PersondataKlientTest.java
@@ -64,6 +64,8 @@ class PersondataKlientTest {
         pdlKlient = new TestPdlKLient(restClient);
     }
 
+    // Bruker ulike metoder for å hente filressurser: File, URL, InputStream
+
     @Test
     void skal_returnere_person() throws IOException {
         // query-eksempel: dokumentoversiktFagsak(fagsak: {fagsakId: "2019186111",
@@ -88,9 +90,9 @@ class PersondataKlientTest {
     }
 
     @Test
-    void skal_returnere_bolk_med_person() throws IOException {
-        var resource = getClass().getClassLoader().getResource("pdl/personBolkResponse.json");
-        var response = DefaultJsonMapper.fromJson(resource.openStream(), HentPersonBolkQueryResponse.class);
+    void skal_returnere_bolk_med_person() throws URISyntaxException {
+        var resource = new File(getClass().getClassLoader().getResource("pdl/personBolkResponse.json").toURI());
+        var response = DefaultJsonMapper.fromJson(resource, HentPersonBolkQueryResponse.class);
         var captor = ArgumentCaptor.forClass(RestRequest.class);
 
         when(restClient.send(captor.capture(), any(Class.class))).thenReturn(response);
@@ -116,8 +118,8 @@ class PersondataKlientTest {
 
     @Test
     void skal_returnere_ident() throws IOException {
-        var resource = getClass().getClassLoader().getResource("pdl/identerResponse.json");
-        var response = DefaultJsonMapper.fromJson(resource.openStream(), HentIdenterQueryResponse.class);
+        var resource = getClass().getClassLoader().getResource("pdl/identerResponse.json").openStream();
+        var response = DefaultJsonMapper.fromJson(resource, HentIdenterQueryResponse.class);
         when(restClient.send(any(RestRequest.class), any())).thenReturn(response);
 
         var queryRequest = new HentIdenterQueryRequest();
@@ -130,9 +132,9 @@ class PersondataKlientTest {
     }
 
     @Test
-    void skal_returnere_bolk_med_identer() throws IOException {
-        var resource = getClass().getClassLoader().getResource("pdl/identerBolkResponse.json");
-        var response = DefaultJsonMapper.fromJson(resource.openStream(), HentIdenterBolkQueryResponse.class);
+    void skal_returnere_bolk_med_identer() throws URISyntaxException {
+        var resource = new File(getClass().getClassLoader().getResource("pdl/identerBolkResponse.json").toURI());
+        var response = DefaultJsonMapper.fromJson(resource, HentIdenterBolkQueryResponse.class);
         when(restClient.send(any(RestRequest.class), any())).thenReturn(response);
 
         var queryRequest = new HentIdenterBolkQueryRequest();
@@ -147,7 +149,7 @@ class PersondataKlientTest {
 
     @Test
     void skal_returnere_ikke_funnet() throws URISyntaxException {
-        var resource = getClass().getClassLoader().getResource("pdl/errorResponse.json");
+        var resource = new File(getClass().getClassLoader().getResource("pdl/errorResponse.json").toURI());
         var response = DefaultJsonMapper.fromJson(new File(resource.toURI()), HentIdenterQueryResponse.class);
         when(restClient.send(any(RestRequest.class), any())).thenReturn(response);
